remove duplicate function, remove commented out code.
