+++ webbreaker/common/logexceptionhelper.py
@@ -27,4 +27,5 @@ class LogExceptionHelper(object):
     def log_info_webinspect_credential_clear_success(self):
         Logger.app.info("Successfully cleared WebInspect credentials from config.ini")
 
-
+    def log_error_invalid_ssl_credentials(self):
+        Logger.app.error("Invalid SSL credentials")
