+++ webbreaker/webinspect/common/loghelper.py
@@ -49,8 +49,5 @@ class WebInspectLogHelper(object):
     def log_error_fetch_webinspect_configs(self):
         Logger.app.error("Fetching webinspect config")
 
-    def log_error_invalid_ssl(self):
-        Logger.app.error("'verify_ssl' must be either 'False' or a full valid CA Path.")
-
     def log_error_not_running_scan(self):
         Logger.app.error("This scan has entered a non-running state, Exiting now!")
